refactor: move dialog setBounds logic to overlay mixin
